refactor: move getMigrationTableName into its own module
getMigrationTableSchema.ts is back to what it is on main, and its spec is split the same way.
