+++ src/utils/getMigrationTableSchema.ts
@@ -1,22 +1,8 @@
-import { getSchemas, quote } from '.';
+import { getSchemas } from '.';
 import type { RunnerOption } from '../runner';
 
 export function getMigrationTableSchema(options: RunnerOption): string {
   return options.migrationsSchema === undefined
     ? getSchemas(options.schema)[0]
     : options.migrationsSchema;
-}
-
-/**
- * The quoted, schema-qualified name of the table storing which migrations have been run.
- *
- * The configured names are used exactly as given: `decamelize` applies to the names used in
- * migrations, not to where their history is kept. So every statement of a run addresses the
- * same table, in a schema spelled like the ones `CREATE SCHEMA` and `SET search_path` use.
- */
-export function getMigrationTableName(
-  options: RunnerOption,
-  schema: string = getMigrationTableSchema(options)
-): string {
-  return `${quote(schema)}.${quote(options.migrationsTable)}`;
 }
